Conditionally install composer

+++ install-standards.sh
@@ -4,19 +4,21 @@
 cd "${HOME}" || exit 1
 
 ### Install Composer
-EXPECTED_SIGNATURE="$(php -r "echo file_get_contents('https://composer.github.io/installer.sig');")"
-php -r "copy('https://getcomposer.org/installer', '$EXPECTED_SIGNATURE.php');"
-ACTUAL_SIGNATURE="$(php -r "echo hash_file('sha384', '$EXPECTED_SIGNATURE.php');")"
+if [ -z "$(which composer)" ]; then
+	EXPECTED_SIGNATURE="$(php -r "echo file_get_contents('https://composer.github.io/installer.sig');")"
+	php -r "copy('https://getcomposer.org/installer', '$EXPECTED_SIGNATURE.php');"
+	ACTUAL_SIGNATURE="$(php -r "echo hash_file('sha384', '$EXPECTED_SIGNATURE.php');")"
 
-if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
-then
-	>&2 echo 'ERROR: Invalid installer signature'
+	if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
+	then
+		>&2 echo 'ERROR: Invalid installer signature'
+		php -r "unlink('$EXPECTED_SIGNATURE.php');"
+		exit 1
+	fi
+
+	php "$EXPECTED_SIGNATURE.php" --install-dir=/usr/local/bin --filename=composer
 	php -r "unlink('$EXPECTED_SIGNATURE.php');"
-	exit 1
 fi
-
-php "$EXPECTED_SIGNATURE.php" --install-dir=/usr/local/bin --filename=composer
-php -r "unlink('$EXPECTED_SIGNATURE.php');"
 
 ### PHPCS
 composer global require -W "dealerdirect/phpcodesniffer-composer-installer:0.7.0" "phpcompatibility/php-compatibility" "wp-coding-standards/wpcs" "automattic/vipwpcs" "woocommerce/woocommerce-sniffs" "automattic/phpcs-neutron-ruleset"
